Refactored выделены структуры, перенесена логика в пакет parser
