+++ README.md
@@ -1,1 +1,5 @@
 # kremlin-parser
+
+```
+CONFIG_PATH=./config/local.yaml go run ./cmd/kremlin-parser/main.go
+```
