+++ go.mod
@@ -4,8 +4,15 @@ go 1.21
 
 require (
 	github.com/gosimple/slug v1.13.1
+	github.com/ilyakaznacheev/cleanenv v1.5.0
 	github.com/spf13/pflag v1.0.5
 	golang.org/x/net v0.20.0
 )
 
-require github.com/gosimple/unidecode v1.0.1 // indirect
+require (
+	github.com/BurntSushi/toml v1.2.1 // indirect
+	github.com/gosimple/unidecode v1.0.1 // indirect
+	github.com/joho/godotenv v1.5.1 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+	olympos.io/encoding/edn v0.0.0-20201019073823-d3554ca0b0a3 // indirect
+)
