+++ internal/entities/Entry/entry.go
@@ -1,0 +1,1 @@
+package Entry
